Use 32-bit context pointer in FUN_00361980

+++ src/Event/mt_evtMain.c
@@ -582,7 +582,7 @@ void FUN_00361140(int param_1,u16 *param_2);
 void FUN_00361200(int param_1,int param_2);
 void FUN_00361270(int param_1);
 void FUN_00361890(int *param_1,u64 param_2);
-void FUN_00361980(u64 param_1);
+void FUN_00361980(int param_1);
 void FUN_00361b60(int *param_1,int param_2);
 void FUN_00361c30(int param_1,int param_2);
 void FUN_00361d60(int param_1,u16 param_2,u32 param_3);
@@ -1597,7 +1597,7 @@ void FUN_00361890(int *param_1,u64 param_2)
 // FUN_00361980 NONMATCHING
 
 
-void FUN_00361980(u64 param_1)
+void FUN_00361980(int param_1)
